requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
oauth2client
httplib2